Fixes issue #224 (Adding new bill generates new unit every time)
Added more case logic to bill creation. If user is guest, and has unit, existing bill and unit gets deleted, and new ones are created. This means guests only have one bill at a time.

Hid the number of occupants from form if user logged in and has a unit.